requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Brotli==1.0.9
cffi==1.15.1
charset-normalizer==2.1.1
commonmark==0.9.1
discord.py==2.0.1
flake8==5.0.4
frozenlist==1.3.1
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
multidict==6.0.2
mypy==0.982
mypy-extensions==0.4.3
orjson==3.8.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycares==4.2.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
rich==12.6.0
tomli==2.0.1
typing_extensions==4.3.0
yarl==1.8.1
